requirements.txt
streamlit
python-docx
doc2pdf # สำหรับแปลงไฟล์ (ถ้า Server รองรับ)
pdf2image
pillow